+++ gui/slinfo.py
