+++ gui/slview.py
